pyproject.toml
[project]
name = "langgraph-for-newbie"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=6.29.5",
    "langchain[anthropic,openai]>=0.3",
    "langgraph>=0.3.34",
    "langsmith>=0.4.8",
    "python-dotenv>=1.1.0",
]
